requirements.txt
Flask==3.0.0
psycopg2-binary>=2.9.0
google-analytics-data==0.18.0
python-dotenv==1.0.0
pandas==2.1.4
schedule==1.2.0
facebook-sdk==3.1.0
requests==2.31.0